refactor(gitgutter): change the way the `CommitHunk` custom function works
First and foremost, we now use the proper gitgutter#hunk#stage function.
Then, there is no input asked anymore : fugitive git commit is launched just
after having staged the hunk.

It is launched with the `-v` flag so that changes are summarized in the
default commit message

Caveat: a hunk being impossible to undo, if the commit operation fails,
the changes are still added to the staging area making it tedious to
cancel