fix(undo): set persistent undo only if home is not encrypted
Persistent undo uses the absolute path as file name so it can become
quite long, pretty quick.
On systems where the $HOME folder is encrypted, there is a ~140 length
limit to filenames that is about to be rapidly hit.
For the moment, we disable persistent undo completely in those
situations